Update ISystem.cs
Mudança de tipo

+++ SegundoProjetoPet/Models/ISystem.cs
@@ -2,7 +2,7 @@
 {
     public interface ISystem
     {
-        ulong Id { get; set; }
+        double Id { get; set; }
 
         string Email { get; set; }
     }
